animated attack scene of enemy

+++ core/src/com/group12/cookiesrising/GameObjects/Enemy.java
@@ -29,7 +29,7 @@ public class Enemy extends AbstractGameObject implements Health,Hittable {
     }
     private String name;
     public double attackPoint;
-    private int speed;
+    private float speed;
     public void setMoney(double money) {
         this.money = money;
     }
@@ -163,11 +163,11 @@ public class Enemy extends AbstractGameObject implements Health,Hittable {
         isHited = true;
     }
 
-    public int getSpeed() {
+    public float getSpeed() {
         return speed;
     }
 
-    public void setSpeed(int speed) {
+    public void setSpeed(float speed) {
         if (speed<=0) speed = 1;
         this.speed = speed;
     }
